data provider trait so we can support snapshot testing?
