+++ crates/sui-graphql-rpc/src/server/mod.rs
@@ -2,5 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod data_provider;
+pub mod json_rpc_data_provider;
 pub mod simple_server;
 mod version;
