+++ crates/sui-graphql-rpc/src/types/protocol_config.rs
@@ -3,7 +3,7 @@
 
 use async_graphql::*;
 
-use crate::server::data_provider::fetch_protocol_config;
+use crate::server::data_provider::DataProvider;
 
 #[derive(Clone, Debug, PartialEq, Eq, SimpleObject)]
 pub(crate) struct ProtocolConfigAttr {
@@ -30,7 +30,8 @@ pub(crate) struct ProtocolConfigs {
 impl ProtocolConfigs {
     async fn configs(&self, ctx: &Context<'_>) -> Result<Option<Vec<ProtocolConfigAttr>>> {
         Ok(Some(
-            fetch_protocol_config(ctx.data_unchecked::<sui_sdk::SuiClient>(), None)
+            ctx.data_unchecked::<Box<dyn DataProvider>>()
+                .fetch_protocol_config(None)
                 .await?
                 .configs,
         ))
@@ -41,18 +42,19 @@ impl ProtocolConfigs {
         ctx: &Context<'_>,
     ) -> Result<Option<Vec<ProtocolConfigFeatureFlag>>> {
         Ok(Some(
-            fetch_protocol_config(ctx.data_unchecked::<sui_sdk::SuiClient>(), None)
+            ctx.data_unchecked::<Box<dyn DataProvider>>()
+                .fetch_protocol_config(None)
                 .await?
                 .feature_flags,
         ))
     }
 
     async fn protocol_version(&self, ctx: &Context<'_>) -> Result<u64> {
-        Ok(
-            fetch_protocol_config(ctx.data_unchecked::<sui_sdk::SuiClient>(), None)
-                .await?
-                .protocol_version,
-        )
+        Ok(ctx
+            .data_unchecked::<Box<dyn DataProvider>>()
+            .fetch_protocol_config(None)
+            .await?
+            .protocol_version)
     }
 
     async fn config(&self, ctx: &Context<'_>, key: String) -> Result<Option<ProtocolConfigAttr>> {
